Remove wpcom_vip_get_page_by_path
Removes the now deprecated `wpcom_vip_get_page_by_path` function and bumps the minimum WordPress vesrion to 6.1.
The alternative is to wrap the function in a `version_compare` statement but as 6.1 is nearly 2 years old, and 5 versions ago and this PR is going into a develop branch a version bump seems logical.
Fixes #134